# docs/openapi.yaml
openapi: 3.0.3
info:
  title: UAV Docking Management System API
  description: |
    Comprehensive REST API for managing UAV operations, real-time tracking, geofencing, and fleet management.
    
    ## Authentication
    The API uses role-based authentication with three user roles:
    - **USER**: Read-only access
    - **OPERATOR**: Can perform UAV operations and updates  
    - **ADMIN**: Full system access including delete operations
    
    ## Rate Limiting
    - Default: 100 requests per minute
    - Admin: 1000 requests per minute
    - API Key: 2000 requests per minute
  version: 1.0.0
  contact:
    name: UAV Management System Team
    email: support@uavmanagement.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.uavmanagement.com
    description: Production server
security:
  - sessionAuth: []
  - apiKeyAuth: []
paths:
...
  /api/geofences:
    get:
      tags:
        - Geofences
      summary: Get all geofences
      description: Retrieve all geofences
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Geofences retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Geofence'
    post:
      tags:
        - Geofences
      summary: Create geofence
      description: Create a new geofence (Admin only)
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeofenceCreateRequest'
      responses:
        '201':
          description: Geofence created successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/geofences/check-point:
    get:
      tags:
        - Geofences
      summary: Check point against geofences
      description: Check if a point violates any geofences
      security:
        - sessionAuth: []
      parameters:
        - name: latitude
          in: query
          required: true
          schema:
            type: number
            format: double
          description: Point latitude
        - name: longitude
          in: query
          required: true
          schema:
            type: number
            format: double
          description: Point longitude
        - name: altitude
          in: query
          schema:
            type: number
            format: double
          description: Point altitude (optional)
      responses:
        '200':
          description: Point check completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeofenceCheckResponse'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Geofence:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Restricted Zone 1"
        description:
          type: string
          example: "No-fly zone around airport"
        fenceType:
          type: string
          enum: [CIRCULAR, POLYGON]
          example: "CIRCULAR"
        boundaryType:
          type: string
          enum: [INCLUSION, EXCLUSION]
          example: "EXCLUSION"
        status:
          type: string
          enum: [ACTIVE, INACTIVE, EXPIRED]
          example: "ACTIVE"
        centerLatitude:
          type: number
          format: double
          example: 40.7589
        centerLongitude:
          type: number
          format: double
          example: -73.9851
        radiusMeters:
          type: number
          format: double
          example: 1000.0
        minAltitudeMeters:
          type: number
          format: double
          example: 0.0
        maxAltitudeMeters:
          type: number
          format: double
          example: 120.0
        priorityLevel:
          type: integer
          example: 1
        violationAction:
          type: string
          example: "RETURN_TO_BASE"
    GeofenceCreateRequest:
      type: object
      required:
        - name
        - fenceType
        - boundaryType
      properties:
        name:
          type: string
          example: "Restricted Zone 1"
        description:
          type: string
          example: "No-fly zone around airport"
        fenceType:
          type: string
          enum: [CIRCULAR, POLYGON]
          example: "CIRCULAR"
        boundaryType:
          type: string
          enum: [INCLUSION, EXCLUSION]
          example: "EXCLUSION"
        centerLatitude:
          type: number
          format: double
          example: 40.7589
        centerLongitude:
          type: number
          format: double
          example: -73.9851
        radiusMeters:
          type: number
          format: double
          example: 1000.0
        minAltitudeMeters:
          type: number
          format: double
          example: 0.0
        maxAltitudeMeters:
          type: number
          format: double
          example: 120.0
        priorityLevel:
          type: integer
          example: 1
        violationAction:
          type: string
          example: "RETURN_TO_BASE"
        notificationEmails:
          type: string
          example: "admin@company.com,operator@company.com"
    GeofenceCheckResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            violations:
              type: array
              items:
                type: object
                properties:
                  geofenceId:
                    type: integer
                    example: 1
                  geofenceName:
                    type: string
                    example: "Restricted Zone 1"
                  violationType:
                    type: string
                    example: "EXCLUSION_VIOLATION"
                  distance:
                    type: number
                    format: double
                    example: 0
                  action:
                    type: string
                    example: "RETURN_TO_BASE"
            isViolation:
              type: boolean
              example: true
            totalViolations:
              type: integer
              example: 1
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "UAV_NOT_FOUND"
            message:
              type: string
              example: "UAV with specified ID not found"
            details:
              type: object
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
  responses:
    BadRequest:
      description: Bad request - invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...